+++ pkg/integrations/buildkite/enterprise/db/module.go
@@ -1,0 +1,7 @@
+package db
+
+import "mash/pkg/di"
+
+func Module(c *di.Container) {
+	c.Register(NewDatabase)
+}
