+++ pkg/integrations/buildkite/enterprise/graphql/module.go
@@ -1,0 +1,7 @@
+package graphql
+
+import "mash/pkg/di"
+
+func Module(c *di.Container) {
+	c.Register(New)
+}
