+++ pkg/integrations/buildkite/enterprise/service/module.go
@@ -1,0 +1,7 @@
+package service
+
+import "mash/pkg/di"
+
+func Module(c *di.Container) {
+	c.Register(New)
+}
