+++ pkg/integrations/buildkite/module/enterprise.go
@@ -1,0 +1,13 @@
+//go:build enterprise
+// +build enterprise
+
+package module
+
+import (
+	"mash/pkg/di"
+	"mash/pkg/integrations/buildkite/enterprise"
+)
+
+func Module(c *di.Container) {
+	c.Import(enterprise.Module)
+}
