+++ pkg/integrations/buildkite/module/oss.go
@@ -1,0 +1,13 @@
+//go:build !enterprise
+// +build !enterprise
+
+package module
+
+import (
+	"mash/pkg/di"
+	"mash/pkg/integrations/buildkite/graphql"
+)
+
+func Module(c *di.Container) {
+	c.Import(graphql.Module)
+}
